+++ sublease-app/src/app/Models/RegistrationRequest/RegistrationRequest.spec.ts
@@ -1,0 +1,7 @@
+import { RegistrationRequest } from './RegistrationRequest';
+
+describe('RegistrationRequest', () => {
+  it('should create an instance', () => {
+    expect(new RegistrationRequest()).toBeTruthy();
+  });
+});
