+++ sublease-app/src/app/Models/RegistrationRequest/RegistrationRequest.ts
@@ -1,0 +1,6 @@
+export class RegistrationRequest {
+  private firstName: string;
+  private lastName: string;
+  private email: string;
+  private password: string;
+}
