+++ sublease-app/src/app/Models/RegistrationResponse/RegistrationResponse.spec.ts
@@ -1,0 +1,7 @@
+import { RegistrationResponse } from './RegistrationResponse';
+
+describe('RegistrationResponse', () => {
+  it('should create an instance', () => {
+    expect(new RegistrationResponse()).toBeTruthy();
+  });
+});
