+++ sublease-app/src/app/Models/RegistrationResponse/RegistrationResponse.ts
@@ -1,0 +1,3 @@
+export class RegistrationResponse {
+  private isSuccess: boolean;
+}
