+++ sublease-app/src/app/app-routing.module.ts
@@ -1,15 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import {Routes, RouterModule} from '@angular/router';
-import { AuthenticateComponent } from './authenticate/authenticate.component';
+import { AuthenticateComponent } from './Components/Forms/authenticate/authenticate.component';
 
 
-const routes:Routes = [
-  {path: "", redirectTo: '/register', pathMatch:'full'},
-  {path: "/register",  component:AuthenticateComponent}
-]
-
-
+const routes: Routes = [
+  {path: '', redirectTo: '/register', pathMatch: 'full'},
+  {path: '/register',  component: AuthenticateComponent}
+];
 
 @NgModule({
   declarations: [],
