+++ sublease-app/src/app/app.module.ts
@@ -1,9 +1,9 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { FormsModule, ReactiveFormsModule }   from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import {HttpClientModule} from '@angular/common/http';
 import { AppComponent } from './app.component';
-import { AuthenticateComponent } from './authenticate/authenticate.component';
+import { AuthenticateComponent } from './Components/Forms/authenticate/authenticate.component';
 import { AppRoutingModule } from './app-routing.module';
 
 @NgModule({
